requirements.txt
asyncio
aiohttp
git+https://git@github.com/lmagyar/python-hass-client.git#egg=hass_client
python-dotenv